Show progress when downloading packages and modify binary package resolution
Protobuild now shows progress when downloading binary packages.  In
addition, it now opts to search recursively for the platform-specific
folder within folders when looking for submodules, instead of generating
a dummy module for binary projects.  This so that when it has two binary
versions for two platforms like so:

* TopLevel
* Package
    * Linux
      * ...
    * Windows
      * ...
* ...

It won't load in definitions from both Linux and Windows (the old way
would place Protobuild.exe in Package; this no longer occurs).  It will
now search Package/Linux for a submodule (when generating for Linux) and
will ignore Package/Windows (unless generating for Windows).