Remove the old module manager GUI
This commit removes the old module manager GUI, which is not maintained and from personal experience, isn't very useful for setting up or configuring projects.  Even when it created the initial project template, it still required you to use a text editor to modify references and add files, something which can be done by reading the documentation instead (and the documentation is much more comprehensive).