Transition xUnit over to a Protobuild package + bug fixes and features
Projects now generate correctly if the framework version is overridden to `v4.5`.
Protobuild now supplies an --upgrade command, which can be used to upgrade packages to the latest version.
Source packages are now always cloned with `git clone` instead of using Git submodules if the parent folder is a Git repo.