Allow package URLs to be redirect either by user config or invocation
Package URLs can be redirected by setting a redirect configuration in
%appdata%\protobuild-redirects.txt (or
~/.config/protobuild-redirects.txt for Linux / Mac).  This configuration
takes the form of "http://original/path -> http://target/path2".
Comments in this file can be made by preceding a line with a hash (#).

Package URLs can also be redirected on the command line by using the
--redirect option, which accepts two arguments, the original and target
URLs for packages.  This command is useful alongside --upgrade-all or
--generate (these redirects will also be passed to submodules).