Duplicate package references in the module hierarchy are now unified
Protobuild will now unify working copies of packages inside the module
hierarchy.  This means that if a parent module depends on a package, and
a submodule depends on the same package, only one instance of that
package will reside in the directory tree, and both modules will refer
to the same copy.

This works for any tree of modules / submodules, including submodules
nested multiple times.