Allow source packages to be added using only their Git URLs
This allows source packages to be added directly using http-git:// and
https-git:// protocols.  When a package has this URL, it will be
detected as a source-only package, with http://... or https://... based
Git URL respectively.  When packages are added with these kinds of URLs,
no package index is queried or required.