Add an --execute option to run the output of a project regardless of it's path
This adds an --execute option which can be used to run the output of a
project without knowing it's path.  This is useful for running tools in
the project that are built as part of the module, such as unit testing
frameworks.  For example, to run the xUnit.NET console runner you could
use:

Protobuild.exe --execute xunit.console assembly_to_test