Experimental MonoDevelop support for referencing SWIG-generated bindings
This adds support for referencing SWIG bindings when a C++ project is
referenced from a C# project.  When the "BindingsGenerator" property on
the C++ project is set to "SWIG", the appropriate build steps are added
to run SWIG and build the C# wrapping DLL.  Then the C# project that
made the reference will have a reference added to the C# wrapping DLL
instead, and the result of the C++ project's build will be added as a
"copy to output on build" file to the root of the C# project.