Fix automatic SWIG installation on Windows + test suite
Fix automatic SWIG installation on Windows due to case-sensitivity.

Fix issue where GetKnownTool XSLT function could return null, which is
not permitted by the Microsoft implementation of XSLT.  Instead of
returning null, an exception is now thrown if the requested tool is not
found and can not be installed.

Fix race conditions in test suite when SWIG is automatically installed,
and fix test for verifying SWIG generator command is correct.

Make test suite show full command when getting an unexpected exit code.