Fix --resolve so it is recursive
This was causing issues with packages nested at a depth of 3 or greater.
Basically when module A depends on package B, which depends on package
C, package C would not be downloaded until package B's project
generation had started.  This would occur after module A loaded
projects, so package C's projects would not appear in the module A
solution file the first time Protobuild was run.