Fix issue where relative paths between submodules was incorrectly calculated
This issue caused binaries to be silently omitted when building binary
packages, which resulted in build failures when binary packages were
used instead of source packages.  The scenario which caused failure was
when a submodule named "SomethingA" was cloned into a folder called
"SomethingB" (i.e. the folder didn't match the submodule name).