Do not initialize a package if the target folder isn't managed by Protobuild
We now prevent Protobuild from initializing packages into their target
folders if the target folder contains at least 1 file or subdirectory,
but doesn't a .pkg file or a .git file or subdirectory.  When these
files are not present, it indicates that something else initialized data
into the target directory prior to Protobuild running and hence there's
no safe way of removing that data to put the package in-place.

This scenario occurs on some build servers like TeamCity, which clone
Git submodules but don't place the .git file that is normally present
for a submodule folder.  Previously Protobuild would delete the copy
that TeamCity placed there and re-clone from the cache.