Set /nodereuse:true for MSBuild on Windows to prevent crashes on build servers
Node reuse in MSBuild is a feature where multiple MSBuild processes will share the same back-end MSBuild nodes on the machine.  However, this means that if one build causes a node to crash, it simultaneously causes other builds dependent on that node to fail.  Since automated build scripts are intended to be used on build servers, turn off node reuse to prevent this scenario from occurring.