Use "git rev-parse" instead of "git status" to workaround hang
When you run "git status" inside a subdirectory of a repository that has Git
LFS enabled, on some machines it results in "git status" hanging as it
continuously runs "git-lfs clean" on each file, despite there being no reason
for the clean filter to be ran at this time.  The issue doesn't hang forever,
but becomes exponentially worse with each new Git LFS file added
to the repository.

To workaround the issue, use "git rev-parse --is-inside-work-tree" instead,
since the use of calling "git status" was to detect whether the current
directory is inside a Git repository.

In addition, this updates Protobuild to indicate when it's running this
command, as previously the execution of "git status" was silent, misleading
the diagnosis of this issue into thinking it was related to the "git fetch" or
"git clone" command that immediately precedes it.