Support creation of NuGet packages from pack
These packages are not usable by the NuGet client, and won't install into Visual Studio, but they can be pushed to
NuGet repositories, which absolves us of having to maintain our own package repository index.  When you push
a NuGet formatted package, it'll attempt to push both a semantic versioned package, and a patched version
of the package with the version set to "0.0.0-GIT<hash>" which will be used by Protobuild to install packages
based on Git version.