Try to use GitHub API first if Git repository is hosted on GitHub
During the installation of Protobuild Manager (hosted on GitHub), we need
to resolve the latest version on NuGet, but we can't do this because Git may not be in the PATH or may
not be available.  We still want developers to be able to install Protobuild Manager without Git on
their PATH (as they may have dedicated shells to use it), so we attempt to use the GitHub API to resolve
the commit hash first.