Add experimental support for custom projects, fix C++ project bugs (#223)
This introduces experimental support for custom projects, which are designed to enable integration with other project generators like CMake and Premake.  In particular, this is intended so that you can import native libraries as a Git submodule or Protobuild package, and have their project generation and build processes integrated with the projects that Protobuild generates.

This commit also fixes some bugs with C++ project generation, and adds a --debug-project-generation option which outputs the intermediate XML files to disk.

C++ projects can now specify additional include paths by adding `NativeInclude` elements in external projects that they reference.  This currently only works for Visual Studio.  For example:

```
<?xml version="1.0" encoding="UTF-8"?>
<ExternalProject Name="netcode.io-extern">
  <Platform Type="Windows">
    <NativeLibraryLink Name="..\netcode.io-import\windows\sodium-release" />
    <NativeInclude Path="..\netcode.io-import\windows" />
    <NativeInclude Path="..\netcode.io-import" />
  </Platform>
</ExternalProject>
```