Fix C++ project generation on macOS (#226)
MonoDevelop / Xamarin Studio completely dropped support for C/C++ projects in their IDE, so the old .cproj format no longer works.  This replaces the format with a generic project (.mdproj) which kicks off an xbuild project as part of CustomCommands, so that the project (and the solution that contains it) can be built with either xbuild or the MonoDevelop IDE, and everything will Just Work(TM).

This doesn't yet fix things for Linux; that will be in a separate PR once I've tested the changes on that platform.