Use Info.plist to detect if the project is a GUI app on macOS
We must use the 'defaults' tool on macOS to determine if the target is a graphical application.  We do this by checking for the presence of NSPrincipalClass in the Info.plist file, which must be present for graphical applications to work.  If it does not exist, we assume it is a command-line application and run it normally like on any other platform.  We have to do this because the native entry point is the only one that has access to the GUI, while the non-native entry point is the only one that has the correct working directory set on startup.