Prefer msbuild on Linux and macOS if available (#242)
* Prefer msbuild on Linux and macOS if available

Microsoft is deprecating xbuild in Mono in favour of MSBuild.  Since Mono 5, using xbuild emits a deprecation warning.  This updates Protobuild to use msbuild on Linux and macOS platforms if available, so that Protobuild will continue working when xbuild is eventually removed.

This commit also removes functional tests relating to the Protobuild package protocol, which has been deprecated for months in favour of NuGet.

* Remove unit test dependent on protobuild.org

* Disable resolution during ModuleInfoGetsUpgradedTest