Ignore exceptions that occur during C++ binding startup (#255)
When C++ bindings start on Windows, the files are copied based on whether the current system is 32-bit or 64-bit. This wraps this process in a generic try/catch, so that exceptions that occur don't cause the program to fail to start.

One example of this is when the same application is started again: the binding DLLs are already in place but currently locked by the first instance of the application, which prevents the second from starting.